+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+* FEATURES
+  * Add [funchook](https://github.com/kubo/funchook) for Python C API functions/methods
+  * Add `fstring` hook
+  * Add `str/bytes/bytearray` `cformat(%)` hook
+
 ## [1.1.3](https://github.com/HXSecurity/DongTai-agent-python/releases/tag/v1.1.3) - 2021-12-03
 
 * FEATURES